1000 inter-route swaps and 500 intra-route swaps
This is a trial run used 1000 swaps between the routes (inter-route swapping) and 500 swaps within each route (intra-route swapping). The results are good but almost none (if any, I didn't check too closely) of the clusters were changed. In fact, most of the csv is unchanged from when I conducted the same trial with no inter-route swapping.